DPT PROJECT/EX-3/requirements.txt
kafka-python==2.0.2
river==0.16.0
joblib==1.2.0
